+++ executor/seqtest/seq_executor_test.go
@@ -1079,3 +1079,83 @@ func (s *seqTestSuite) TestMaxDeltaSchemaCount(c *C) {
 	c.Assert(variable.GetMaxDeltaSchemaCount(), Equals, int64(2048))
 	tk.MustQuery("select @@global.tidb_max_delta_schema_count").Check(testkit.Rows("2048"))
 }
+
+func (s *seqTestSuite) TestBatchDML(c *C) {
+	tk := testkit.NewTestKit(c, s.store)
+	tk.MustExec("use test")
+	tk.MustExec("create table t (i int key, j varchar(20))")
+	orgConf := config.GetGlobalConfig()
+	originLimit := atomic.LoadUint64(&kv.TxnEntryCountLimit)
+	defer func() {
+		config.StoreGlobalConfig(orgConf)
+		atomic.StoreUint64(&kv.TxnEntryCountLimit, originLimit)
+	}()
+	atomic.StoreUint64(&kv.TxnEntryCountLimit, 2)
+	newConf := config.NewConfig()
+	newConf.EnableBatchDML = true
+	config.StoreGlobalConfig(newConf)
+	tk.MustExec("set @@session.tidb_dml_batch_size=1;")
+	tk.MustExec("set @@session.tidb_disable_txn_auto_retry=0;")
+
+	// Test auto commit transaction.
+	tk.MustExec("insert into t values (1, 'a'), (2, 'b'), (3, 'c')")
+	tk.MustQuery("select * from t order by i").Check(testkit.Rows("1 a", "2 b", "3 c"))
+	tk.MustExec("replace into t values (1, '-a'), (2, '-b'), (3, '-c')")
+	tk.MustQuery("select * from t order by i").Check(testkit.Rows("1 -a", "2 -b", "3 -c"))
+	tk.MustExec("update t set i = -i")
+	tk.MustQuery("select * from t order by i desc").Check(testkit.Rows("-1 -a", "-2 -b", "-3 -c"))
+	tk.MustExec("delete from t")
+	tk.MustQuery("select * from t").Check(testkit.Rows())
+
+	// Test transaction block.
+	tk.MustExec("begin")
+	tk.MustExec("insert into t values (1, 'a'), (2, 'b'), (3, 'c')")
+	tk.MustQuery("select * from t order by i").Check(testkit.Rows("1 a", "2 b", "3 c"))
+	tk.MustExec("replace into t values (1, '-a'), (2, '-b'), (3, '-c')")
+	tk.MustQuery("select * from t order by i").Check(testkit.Rows("1 -a", "2 -b", "3 -c"))
+	tk.MustExec("update t set i = -i")
+	tk.MustQuery("select * from t order by i desc").Check(testkit.Rows("-1 -a", "-2 -b", "-3 -c"))
+	tk.MustExec("delete from t")
+	tk.MustQuery("select * from t").Check(testkit.Rows())
+	tk.MustExec("commit")
+
+	// Test disable batch DML.
+	config.GetGlobalConfig().EnableBatchDML = false
+	_, err := tk.Exec("insert into t values (1, 'a'), (2, 'b'), (3, 'c')")
+	c.Assert(kv.ErrTxnTooLarge.Equal(err), IsTrue)
+	tk.MustQuery("select * from t").Check(testkit.Rows())
+	_, err = tk.Exec("replace into t values (1, '-a'), (2, '-b'), (3, '-c')")
+	c.Assert(kv.ErrTxnTooLarge.Equal(err), IsTrue)
+	tk.MustQuery("select * from t").Check(testkit.Rows())
+	tk.MustExec("insert into t values (1, 'a')")
+	tk.MustExec("insert into t values (2, 'b')")
+	tk.MustExec("insert into t values (3, 'c')")
+	_, err = tk.Exec("update t set i = -i")
+	c.Assert(kv.ErrTxnTooLarge.Equal(err), IsTrue)
+	tk.MustQuery("select * from t order by i").Check(testkit.Rows("1 a", "2 b", "3 c"))
+	_, err = tk.Exec("delete from t")
+	c.Assert(kv.ErrTxnTooLarge.Equal(err), IsTrue)
+	tk.MustQuery("select * from t order by i").Check(testkit.Rows("1 a", "2 b", "3 c"))
+
+	// Test batched retry
+	config.GetGlobalConfig().EnableBatchDML = true
+	atomic.StoreUint64(&kv.TxnEntryCountLimit, 10)
+	tk1 := testkit.NewTestKit(c, s.store)
+	tk1.MustExec("use test")
+	tk1.MustExec("set @@session.tidb_disable_txn_auto_retry=0;")
+	tk.MustExec("set @@session.tidb_dml_batch_size=2;")
+
+	tk.MustExec("begin")
+	tk1.MustExec("update t set j = 'e'")
+	_, err = tk.Exec("update t set j = 'd'")
+	c.Assert(executor.ErrBatchDMLFail.Equal(err), IsTrue, Commentf("error %s", err))
+	tk.MustQuery("select * from t order by i").Check(testkit.Rows("1 e", "2 e", "3 e"))
+
+	// Test normal retry
+	config.GetGlobalConfig().EnableBatchDML = false
+	tk.MustExec("begin")
+	tk1.MustExec("update t set j = 'f' where i = 1")
+	tk.MustExec("update t set j = 'd' where i = 1")
+	tk.MustExec("commit")
+	tk.MustQuery("select * from t order by i").Check(testkit.Rows("1 d", "2 e", "3 e"))
+}
